factored out the oregon scientific support and re-added acurite rain gauge support
