+++ include/rtl_433_devices.h
@@ -10,5 +10,7 @@ extern r_device elv_em1000;
 extern r_device elv_ws2000;
 extern r_device lacrossetx;
 extern r_device acurite5n1;
+extern r_device acurite_rain_gauge;
+extern r_device oregon_scientific;
 
 #endif /* INCLUDE_RTL_433_DEVICES_H_ */
